+++ CarCareManager/composeApp/src/commonMain/kotlin/com/echcoding/carcaremanager/domain/repository/VehicleRepository.kt
@@ -10,10 +10,10 @@ import kotlinx.coroutines.flow.Flow
  */
 interface VehicleRepository {
     suspend fun upsertVehicle(vehicle: Vehicle)
-    suspend fun deleteVehicleById(id: Long)
-    suspend fun getVehicleById(id: Long): Vehicle?
+    suspend fun deleteVehicleById(id: Int)
+    suspend fun getVehicleById(id: Int): Vehicle?
     suspend fun getAllVehicles(): Flow<List<Vehicle>>
 
     fun getActiveVehicle(): Flow<Vehicle?>
-    suspend fun setActiveVehicle(vehicleId: Long)
+    suspend fun setActiveVehicle(vehicleId: Int)
 }
