+++ CarCareManager/composeApp/src/commonMain/kotlin/com/echcoding/carcaremanager/presentation/vehicle/vehicle_list/VehicleListAction.kt
@@ -7,10 +7,10 @@ import com.echcoding.carcaremanager.domain.model.Vehicle
  * The intent represents an user action or event that triggers a state change
  */
 sealed interface VehicleListAction {
-    data class OnSelectVehicleClick(val vehicleId: Long) : VehicleListAction
+    data class OnSelectVehicleClick(val vehicleId: Int) : VehicleListAction
     data object OnAddVehicleClick : VehicleListAction
     data class OnEditVehicle(val vehicle: Vehicle) : VehicleListAction
-    data class OnDeleteVehicle(val vehicleId: Long) : VehicleListAction
+    data class OnDeleteVehicle(val vehicleId: Int) : VehicleListAction
     data object OnConfirmDeleteVehicle : VehicleListAction
     data object OnDismissDeleteDialog: VehicleListAction
     data class OnTabSelected(val index: Int) : VehicleListAction
